requirements.txt
aioredis
black
celery
channels
channels-redis
crispy-bootstrap5
dateparser
python-decouple
Django
django-filter
django-crispy-forms
django-extensions
django-js-asset
django-mptt
django-simple-history
whitenoise
hiredis
Jinja2
#lxml
pipdeptree
psycopg
python-dateutil
pytz
redis
uvicorn
